requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
httpx==0.25.2
openai>=1.6.1,<2.0.0
langchain==0.0.340
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers>=2.3.0
pypdf2==3.0.1
python-docx==1.1.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
jinja2==3.1.2
aiofiles==23.2.0
pydantic>=2.0.0,<3.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv>=1.0.0 